merge master

+++ solr/core/src/test/org/apache/solr/cloud/TestDistributedMap.java
@@ -35,9 +35,9 @@ public class TestDistributedMap extends SolrTestCaseJ4 {
   protected static ZkTestServer zkServer;
   
   @BeforeClass
-  public static void setUpClass() throws InterruptedException {
+  public static void setUpClass() throws Exception {
     zkDir = createTempDir("TestDistributedMap");
-    zkServer = new ZkTestServer(zkDir.toFile().getAbsolutePath());
+    zkServer = new ZkTestServer(zkDir);
     zkServer.run();
   }
   
@@ -48,8 +48,10 @@ public class TestDistributedMap extends SolrTestCaseJ4 {
       zkServer.shutdown();
       zkServer = null;
     }
-    FileUtils.deleteDirectory(zkDir.toFile());
-    zkDir = null;
+    if (null != zkDir) {
+      FileUtils.deleteDirectory(zkDir.toFile());
+      zkDir = null;
+    }
   }
   
   public void testPut() throws KeeperException, InterruptedException {
@@ -171,7 +173,7 @@ public class TestDistributedMap extends SolrTestCaseJ4 {
   }
   
   protected String getAndMakeInitialPath(SolrZkClient zkClient) throws KeeperException, InterruptedException {
-    String path = String.format(Locale.ROOT, "/%s/%s", getClass().getName(), getTestName());
+    String path = String.format(Locale.ROOT, "/%s/%s", getClass().getName(), getSaferTestName());
     zkClient.makePath(path, false, true);
     return path;
   }
